requirements.txt
streamlit
pillow
pandas
joblib
pypdf
openpyxl
xlsxwriter
scikit-learn==1.3.2
numpy
